[Kush] BQ label longer than 63 characters will be trimmed
